+++ webapp/test/equip.test.js
@@ -3,7 +3,6 @@
 const test = require('node:test');
 const assert = require('node:assert');
 const fs = require('node:fs');
-const os = require('node:os');
 const path = require('node:path');
 
 const paths = require('../services/pathService');